Problem statement: Make the last access time reset feature in windows local crawl configurable so that it can be turned on/ off as per customer environment.
Solution:
1. Added a boolean "lastAccessTimeResetFlag" in WindowsFileSystemType class and
made it configurable through spring configuration file. 
2. By default it is kept "false" to avoid extra file I/O
3. The above mentioned boolean is then passed to "WindowsReadOnlyFile" which
based on the value of the flag decides whether or not to get the last access
time of the file. 
4. WindowsReadOnlyFile then passes the boolean flag to "WindowsFileInputStream"
class which decides whether or not to set the time back to the file based on the
value of the boolean.

Testing:
Tested it manually for "true" as well as "false" case and it is working as
expected