Issue 4568074: Change byte array size for MimeTypeFinder
(Regardless whether we keep using this class or use CMs util i would like to
check in this CL in FS connector)

MimeTypeFinder.getBytes() has an internal buffer that
it uses to read a byte[] of content for the File Connector.
This buffer is used to try to determine the mime type by
magic number.  Unfortunately, the buffer is just a bit too
small.  At 2056 bytes, the buffer is not large enough to
match the the following MagicMime ids:

2080  string  Documento\ Microsoft\ Word\ 6 application/msword
2080  string  Microsoft\ Excel\ 5.0\ Worksheet  application/vnd.ms-excel
2080  string  Microsoft\ Word\ 6.0\ Document  application/msword
2112  string  MSWordDoc     application/msword
2114  string  Biff5         application/vnd.ms-excel
32769    string    CD001     application/x-iso9660-image

(From mime-util's magic.mime resource.  First number is
byte offset into the file to look for the magic value.)

We don't need 32K to identify a ISO 9660 CD-ROM image,
but we should have at least a 2120 byte buffer to ID the
Microsoft Office docs.  I suggest you increase the buffer
size to 2560 bytes (5 x 512 byte blocks).

CL URL: http://codereview.appspot.com/4568074/