Partial fix for bug 6679622: Include createTime in lastModified calculation for SMB and Windows file systems.
According to this Microsoft document: http://support.microsoft.com/kb/299648,
moving or renaming a file within the same file system does not change either 
the last-modify timestamp of a file or the create timestamp of a file.

However, copying a file or moving it across filesystems (which involves an
implicit copy) sets a new create timestamp, but does not alter the last
modified timestamp.

This change set enhances SmbReadonlyFile.getLastModified() and 
WindowsReadonlyfile.getLastModified() to return the newer of create timestamp
and last modified timestamp.

This will detect copying of files and moves across filesystems, but not
moves or renames within a filesystem. The latter  might be detected via a
change in the last modified timestamp of the parent directory for the file,
but this would require extensive reworking of FileIterator.

Code Review:
http://codereview.appspot.com/6333050/