Fix Bug 6977695: Filesystem traversal restarts after RepositoryExceptions.
RepositoryExceptions would bail out of the Traverser thread that
was traversing a startpoint.  The FileLister would then wait
the error wait and start the Traverser over at the startpoint.

This might cause a Traverser to never successfully traverse
all files/directories from a startpoint if there are suffienctly
many that a transient repository access error might occur.
Any transient error would force the traversal to restart at
the beginning.

This change pushes RepositoryExecption handling into the
Traverser.  This allows the Traverser to sleep for the
error wait and attempt to resume where it left off.

In a related issue, FileIterator.hasNext() was throwing
RepositoryDocumentExceptions. However the call to hasNext()
was outside the try/catch block for RepositoryDocumentExceptions,
so the RDE would throw out of the traverse() method (and get
handled as a RepositoryException), rather than skip the 
document and go onto the next.  This would terminate
the traverser, and force a new traverser to start over
at the startpoint.

This change also handles failure to get the share ACL
differently.  In this case, the Traverser does throw
a RepositoryException, rather than feed a bunch of 
documents whose inherited ACLs cannot be resolved.

TODO: Because of the design of FileIterator, it is not
possible to retry the failed document if next() throws
the RepositoryException, so this implementation skips
the document that encountered the problem (like
RepositoryDocumentException).

Code Review:
http://codereview.appspot.com/6464067