This change addresses numerous problems with SmbAclBuilder
See bugs b/7307947 and b/7309510.

1) SmbAclBuilder.checkAndAddAces() would return prematurely if it
   was skipping inherited ACEs and encountered one.  The return
   statement there should have been a continue statement.

2) The log messages in checkAndAddAces and checkAndAddParentInheritOnlyAces
   were really misleading.  Partly because the text was just wrong.
   And partly because SmbAclBuilder made a couple of passes over the
   file's array of ACEs - one passing true to checkAndAddAces, and
   another passing false.  This would log messages where inherited
   ACEs were filtered out, then later log those same ACEs as added.
   The contradictory messages could be confusing to the user.

3) Inherit-only ACEs were being fed for the share/directory for
   which they were set, when they should only be supplied for
   the immediate children of that node, not the node itself.

4) The tests were lacking, to say the least.


In the case of problem 1) (the return instead of continue), 
the following, nearly identical checkAndAddParentInheritOnlyAces
method got it right.  When I found myself trying to fix 3)
and creating still another nearly identical method, I decided
to have checkAndAddAces take a Predicate to determine whether
to add the ACE to the lists of ACEs.  I converted the
isInherited* methods to Predicates, and was able to remove
several redundant and superfluous methods from SmbAclBuilder.
As a result of the refactoring, the confusing log messages
were removed.

Finally I enhanced SmbAclBuilerTest to actually test
the fixes for 1) and 3).  SmbAclBuilderTest also had
redundant helper methods for determining whether a 
Collection of principals contained one with a specific 
name.  I eliminated one of them.

Code Review:
http://codereview.appspot.com/6619055