Fix Issue 7291526 Part 1: Do not send inherit-from parent if file inherits no ACLs
If a file/directory has no INHERITED ACEs in its ACEs, then do not send
ACL 'inherit-from' pointing at the parent.  It should 'inherit-from' the
share (like the startpoint).

Similarly, do not bother running the parent for INHERIT_ONLY ACEs,
because they will not apply to that child.

This does not address the reported problem DIRECT ACEs on the file
not getting fed, as we have not been able to reproduce that case yet.

Code Review:
https://codereview.appspot.com/6643063/