Synchronize access to the JCIFS library.
This change, for an experimental build, synchronizes access to the
JCIFS package via a big, giant lock, implemented in SmbFileDelegate.

This is to test whether a specific customer's NT_STATUS_NOT_FOUND
errors and "The parameter is incorrect" errors are rooted in the
poor thread-safety of the JCIFS library.

Code Review:
http://codereview.appspot.com/6843053