Fix Issues 7690736, 7684717: Incorrect ACL inheritance.
Issue 7690736: Inherit-only ACEs should appear in an ACL associated with the
parent directory, not the child.  We were folding Inherit-only ACEs from the 
parent onto the child.  Consequently, if the parent ACE changed, that change
would not be reflected in the child's copy.

Issue 7684717: Files incorrectly inherit list folder contents permission.
Microsoft Windows inherited permissions may differ between child directories
and child files.  Specifically, ACEs can be marked to be inherited by files,
containers, or both.  Consequently, list folder contents permission grants
read access to folders and sub-folders, but not the files contained within 
them.

To handle the latter issue, we create two separate ACL documents for 
each directory we are feeding: one that can be an inherit-from for files,
and another that is the inherit-from for sudirectories.  I added a factory
method to FileDocument that can take a ReadonlyFile and return one or more
FileDocuments representing that file.  This allows FileLister to feed two
separate ACL documents for each directory.

The File System Connector does not consider directories to be indexable 
Documents, and feeds them only as ACL records.  Consequently, accurately
modelling the permissions on the directories themselves is not really
needed; as long as the permissions eventually inherited by regular files
are correct.  

This provides an easy fix for the Inherit-Only permissions, which are 
supposed to apply to descendants, but not the directory itself.  Since
we don't actually care of the directory's permissions are correct, we
can safely ignore the Inherit-Only rule and include that permission in
the directory's inheritable ACLs.  

No-Propagate permissions can be handled similarly.  They are supposed to 
apply to the immediate children of a directory, but not propagate further.
Since we now have separate ACLs that are inherited by child directories and
child files, No-Propagate is easy to model as well.  The No-Propagate 
permissions are included in the ACLs inherited by files (which can propagate
them no further), but not included in the ACLs inherited by directories
(which should not propagate them to any of thier children).

Code Review:
http://codereview.appspot.com/7304049