Fix Issue 6444005: All pipe instances are busy.
The problem here is that this error intermittently strips the ACL
from a document when we send the document in a feed again.  

By stripping the ACL, authorization for the document falls back to
Connector AuthZ, which will only work if an external AuthN mechanism
is in place.

I really cannot do much about the "All pipe instances are busy"
conditions, but I can change how these intermittent ACL errors
are handled.

This change adds an advanced configuration option, "useAuthzOnAclError".
If this option is "true", then no ACL will be provided, forcing
Connector AuthZ to be used.  If the option is "false" (the default),
then a RepositoryDocumentException will be thrown, forcing the 
document to be skipped.

Code Review:
http://codereview.appspot.com/12487044